stop advertising the running of tests to end users
I've realized that this was broken. (The tests were not installed and could be
only run for inplace-builds linked from somewhere else.)  Fixing this would be
somewhat hacky.  I think now that it was a bad idea in the first place to
advertise the running of unit tests to end users.