add explicit documentation on lead numbering
It was not previously made explicit anywhere in the docs on how
the leads were numbered. This commit adds a line to the docstring
of `kwant.builder.Builder.attach_lead` which explains the meaning
of the return value. It also adds a few lines to the first tutorial
explaining the numbering.