setup.py: disable cython by default
The pip tool messes up the timestamps of files, so with our previous strategy
of enabling cython by default the build fails if Cython is not available.

Also, turn the "warning" into a "caution" message that is less frightening and
send it to stdout only, so that it does not appear with pip.