allow systems to accept named parameters
Hamiltonian value functions may now have signatures that depend on
different parameters. On finalization, Builders now inspect all of their
value functions and store the names of the parameters on which the value
function depends. Then, when evaluating the Hamiltonian, a particular
value function is only passed the parameters on which it depends.

'System.hamiltonian' has been updated to accept an extra keyword-only
parameter, 'params', which is a dictionary mapping parameter names to
values. This is mutually exclusive with the existing '*args'. All
top-level API that takes an 'args' parameter now also takes a
keyword-only parameter, 'params'