change TranslationalSymmetry.act() to reject non-integer input
Previously, if a tuple of floats was passed as the 'domain' argument to
'act' then the tags of the resulting sites would be floats, rather than
integers. This is because the tags of sites created inside 'act' were
not normalized to integers, as they would usually be, and this causes
bugs in other locations where an integer dtype is assumed.