make Builder.fill raise exception when no new sites are added
Previously, 'syst.fill(template, shape, start)' would return
successfully even if all sites in domain 'start' are outside of 'shape'.
This is in contrast to the behaviour of 'lattice.shape'.
In addition, if the target builder already has sites in the start domain
and `overwrite=False` then we should raise an exception, as this is
probably not what the user expects.

Closes #107