CI: separate conda packaging into separate jobs
We previously restricted conda package uploading to master branch on the
main Kwant repository, as otherwise this step would fail for forked
repositories without a valid anaconda key. Now we build the conda
packages and upload to anaconda in two separate jobs, and the upload job
is only run on the main Kwant repository.