add tests for default parameter values in value functions
The semantics for default parameters are as follows. If a value
function has a parameter that takes a default value, then an
exception is raised if the user ever tries to assign a value
to this parameter via 'params'. These semantics are chosen
to eliminate the possibility that "forgotten" default parameters
are not silently overwritten.