kpm: refactor the SpectralDensity API
It is now impossible to reduce the number of random vectors or moments.
It is also no longer possible to set the number of energy points used by
the method -- this can be supplied by providing a sequence of energies
when calling the KPM object. We replace the 'increase_energy_resolution'
and 'increase_accuracy' methods by 'add_moments' and 'add_vectors'.
The tests have also been updated to reflect these changes.