make args no longer keyword-only
The purpose of keyword-only args is to allow functions that take a
variable number of args but also a variable number of "options" without
having to resort to (*args, **kwargs).  Builder value functions do not
fall into this category.  Hence, there is no reason to pay the price (in
terms of flexibility and efficiency) for not much gain.