improve the tutorial script and image buiding machinery
The basic idea remains the same, but should work much better now.  The
difference is that images/example.py.diff is now a patch with complete
context and becomes the authoritative source for both the visible flavor
of an example as well as for its figure-generating variant.  Both
tutorial/example.py and images/example.py are extracted from this file
by 'make html'.  Thanks to the complete context the diffs are quite
readable and may be modified directly.

Alternatively, one may also modify the generated scripts.  When
tutorial/example.py has been modified, it will be patched and saved as
images/example.py.  The patching is done using the tool 'wiggle' that
works much better than 'patch'.  If a conflict occurs, conflict markers
are added to the output file and its dated back to the dawn of time
(i.e. 1970) in order to mark the conflicts as not yet resolved.  After
resolving, 'make html' is simply run again.

Upon a successful execution of the figure-generating script the diff
gets recreated automatically, 'generate-diffs' no longer exists.