move tutorial script generation, reorganize doc/source
I assume that most people encounter the tutorial example scripts by
reading the documentation, and not by viewing the tutorial subdirectory
of a checked-out Kwant source.

That's the motivation for moving all the manipulations of tutorial
scripts from setup.py to doc/Makefile.  Previously, a successful 'make
html' would require a preceding execution of 'setup.py build_tut'.  Now,
a simple 'make html' is enough.

While at it, I reorganized where the example scripts and their outputs
are stored.  Everything is now in 'doc/source/code' with its three
subdirectories 'download' (to be shown to readers), 'include' (with
include markers), and 'figure' (figure generation & figures).  This
organization is clearer and also makes the generation of figures
separate from the tutorial.