fix confusing warning about deprecated argument grid_spacing
The warning used to be emitted from build_discretized().  Whenever
discretize(), another API function that internally calls
build_discreitzed() would be called with the deprecated argument, a
warning would appear for a line in the code of discretize().

This has been fixed by factoring out the check, and calling it from both
functions.