improve heuristic for setting colormap limits in plotter.map
Previously we used the matplotlib default behavior of setting
the colormap limits (vmin and vmax) to the limits of the plotted
data, if they are not provided by the user. Now we set the
colormap limits to the 2nd and 98th percentile of the input
data, and stretch the limits by 10% on either end in an attempt
to include all the data in the limits. If we cannot do so
we set the limits to the percentiles plus the stretch
and issue a warning.

Also, whenever the data falls outside the colorbar limits
(even when the latter are set by the user) we set the appropriate
ends of the colorbar to be pointy, indicating that the data
extends beyond the colorbar.

Closes #183