Fix compatibility issues with Sympy 1.2. Closes #203.
SymPy 1.2 brought changes to LambdaPrinter that broke kwant.continuum
for inputs containing functions like "sin", "cos", or constant "pi".
By this fix we assert that "kwant.continuum.sympify" represents in
its output all functions as AppliedUndef in favour to built-in SymPy
functions. We also put a hard constaint that all functions and constants
in the input expression (with except to pi) must be provided through params.