skip qsymm tests if sympy and qsymm are not installed
This is a partial application of the changeset proposed in
https://gitlab.kwant-project.org/kwant/kwant/merge_requests/274/,
which removes the need for a 'conftest.py', instead using
'pytest.importorskip'.

We apply this change only to 'qsymm', as applying this change elsewhere
requires rethinking where tests are put in the package structure of
Kwant. This minimal change allows us to build packages for Conda and
Debian, which is necessary for releasing Kwant 1.4.