fix avoidable backwards incompatibility: System.parameters
When System.parameters was introduced, we thought it would be a good idea to
also use this new attribute.  The plan was to check in hamiltonian_submatrix()
whether any unneeded parameters are provided by the user.  This turned out to
be impossible, so the only check that remained was whether all the necessary
parameters are indeed provided.  However, this check is not necessary, because
the 'hamiltonian' method is in any way obliged to signal an error when it does
not obtain the necessary parameters.