Use correct bloch basis in qsymm wrapper.
The function which finds the symmetries of kwant builders does not find all real space symmetries if the unit cell contains more than one sites, because the conversion from builder to qsymm model does not use a basis which preserves information on the real space position of sites. The tests do not catch this, because they only have one site per unit cell. This fixes the basis choice, and adds a test.