update all relevant 'isinstance' checks on systems
We now have *System and *System2, so the old 'isinstance' checks
will not be complete. We add convenience functions in 'kwant.system'
and 'kwant.builder' to check against all the relvant classes.
We leave the original 'isinstance' checks in 'kwant.physics.gauge'
because that module does not work with vectorize systems yet.