move infinite system docstring from mixin to abstract base class
The docstring refers to site orderings, and explicit references to
the graph; these are things that vectorized systems keep for
backwards compatibility, but we wish to move away from. For this
reason we move the docstring to the (unvectorized) infinite system
class and add a more specific docstring to the vectorize infinite
system class.