update libraries for linking against Mumps in conda-forge
Recently conda-forge updated their packaging so that sequential
Mumps is a separate package,
[mumps-seq](https://anaconda.org/conda-forge/mumps-seq),
which names the libraries with a suffix '_seq' (contrary to how
Debian does it).
In addition we no longer need to explicitly link against Scotch,
pord or gfortran as the conda-forge built mumps libraries already
do this.
All tests still pass without explicitly linking against mpiseq,
so it is left out.