remove code for specially treating umfpack solver
We relied on accessing private namespace in scipy, which is bound to break and
emits deprecation warnings. Additionally I verified that there is no way for a
user to obtain a functioning scikit-umfpack aside of building it from source.
This makes the additional effort of implementing an own solver negligible.